attempt to improve crackling in audio
the issue here was that the destination buffer was using the same length as the
source buffer and the destination buffer needs to be longer or shorter based on
ratio between the input and output sample rates. This is still not a perfect
solution, and needs to be revisited for further improvement.